feat: create a refresh method in the DOM object and correct finding the index of the book in the books array in removeBook method
the refresh method is called whenever the DOM needs to be refreshed - that is when a new book is added, the book status is changed or the book is removed
the removeBook function before removed all the books from the books array, this was bad behaviour. Using indexOf returned undefined.